+++ src/core/StackElement.h
@@ -10,16 +10,16 @@
 #define MSK_STACK_ELEMENT_TYPE_PROPERTIES						(STACK_ELEMENT_PROPERTY_MAX-1)
 #define STACK_ELEMENT_TYPE_PROPERTIES(prop)						((prop)&MSK_STACK_ELEMENT_TYPE_PROPERTIES)
 
-#define STACK_ELEMENT_IS_OBJECT_SCRIPT_OBJECT(stk) 				(((stk)->properties & zetscript::STACK_ELEMENT_PROPERTY_OBJECT) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_OBJECT_SCRIPT_OBJECT))
-#define STACK_ELEMENT_IS_ITERATOR_OBJECT_SCRIPT_OBJECT(stk)		(((stk)->properties & zetscript::STACK_ELEMENT_PROPERTY_OBJECT) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_OBJECT_ITERATOR_SCRIPT_OBJECT))
-#define STACK_ELEMENT_IS_STRING_SCRIPT_OBJECT(stk) 				(((stk)->properties & zetscript::STACK_ELEMENT_PROPERTY_OBJECT) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_STRING_SCRIPT_OBJECT))
-#define STACK_ELEMENT_IS_STRING_ITERATOR_SCRIPT_OBJECT(stk)		(((stk)->properties & zetscript::STACK_ELEMENT_PROPERTY_OBJECT) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_STRING_ITERATOR_SCRIPT_OBJECT))
-#define STACK_ELEMENT_IS_ARRAY_SCRIPT_OBJECT(stk) 				(((stk)->properties & zetscript::STACK_ELEMENT_PROPERTY_OBJECT) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_ARRAY_SCRIPT_OBJECT))
-#define STACK_ELEMENT_IS_ITERATOR_ARRAY_SCRIPT_OBJECT(stk)		(((stk)->properties & zetscript::STACK_ELEMENT_PROPERTY_OBJECT) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_ARRAY_ITERATOR_SCRIPT_OBJECT))
-#define STACK_ELEMENT_IS_CLASS_OBJECT(stk) 						(((stk)->properties & zetscript::STACK_ELEMENT_PROPERTY_OBJECT) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id>=zetscript::SCRIPT_TYPE_ID_CLASS_SCRIPT_OBJECT))
-#define STACK_ELEMENT_IS_VAR_REF_OBJECT(stk) 					(((stk)->properties & zetscript::STACK_ELEMENT_PROPERTY_OBJECT) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_VAR_REF_SCRIPT_OBJECT))
-#define STACK_ELEMENT_IS_MEMBER_FUNCTION_OBJECT(stk)			(((stk)->properties & zetscript::STACK_ELEMENT_PROPERTY_OBJECT) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_MEMBER_FUNCTION_SCRIPT_OBJECT))
-#define STACK_ELEMENT_IS_ZETOBJECT_SCRIPT_OBJECT(stk) 			(((stk)->properties & zetscript::STACK_ELEMENT_PROPERTY_OBJECT) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_ZETSCRIPT_SCRIPT_OBJECT))
+#define STACK_ELEMENT_IS_OBJECT_SCRIPT_OBJECT(stk) 				(((stk)->properties & (zetscript::STACK_ELEMENT_PROPERTY_OBJECT|zetscript::STACK_ELEMENT_PROPERTY_OBJECT_IN_CONTAINER)) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_OBJECT_SCRIPT_OBJECT))
+#define STACK_ELEMENT_IS_ITERATOR_OBJECT_SCRIPT_OBJECT(stk)		(((stk)->properties & (zetscript::STACK_ELEMENT_PROPERTY_OBJECT|zetscript::STACK_ELEMENT_PROPERTY_OBJECT_IN_CONTAINER)) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_OBJECT_ITERATOR_SCRIPT_OBJECT))
+#define STACK_ELEMENT_IS_STRING_SCRIPT_OBJECT(stk) 				(((stk)->properties & (zetscript::STACK_ELEMENT_PROPERTY_OBJECT|zetscript::STACK_ELEMENT_PROPERTY_OBJECT_IN_CONTAINER)) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_STRING_SCRIPT_OBJECT))
+#define STACK_ELEMENT_IS_STRING_ITERATOR_SCRIPT_OBJECT(stk)		(((stk)->properties & (zetscript::STACK_ELEMENT_PROPERTY_OBJECT|zetscript::STACK_ELEMENT_PROPERTY_OBJECT_IN_CONTAINER)) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_STRING_ITERATOR_SCRIPT_OBJECT))
+#define STACK_ELEMENT_IS_ARRAY_SCRIPT_OBJECT(stk) 				(((stk)->properties & (zetscript::STACK_ELEMENT_PROPERTY_OBJECT|zetscript::STACK_ELEMENT_PROPERTY_OBJECT_IN_CONTAINER)) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_ARRAY_SCRIPT_OBJECT))
+#define STACK_ELEMENT_IS_ITERATOR_ARRAY_SCRIPT_OBJECT(stk)		(((stk)->properties & (zetscript::STACK_ELEMENT_PROPERTY_OBJECT|zetscript::STACK_ELEMENT_PROPERTY_OBJECT_IN_CONTAINER)) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_ARRAY_ITERATOR_SCRIPT_OBJECT))
+#define STACK_ELEMENT_IS_CLASS_OBJECT(stk) 						(((stk)->properties & (zetscript::STACK_ELEMENT_PROPERTY_OBJECT|zetscript::STACK_ELEMENT_PROPERTY_OBJECT_IN_CONTAINER)) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id>=zetscript::SCRIPT_TYPE_ID_CLASS_SCRIPT_OBJECT))
+#define STACK_ELEMENT_IS_VAR_REF_OBJECT(stk) 					(((stk)->properties & (zetscript::STACK_ELEMENT_PROPERTY_OBJECT|zetscript::STACK_ELEMENT_PROPERTY_OBJECT_IN_CONTAINER)) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_VAR_REF_SCRIPT_OBJECT))
+#define STACK_ELEMENT_IS_MEMBER_FUNCTION_OBJECT(stk)			(((stk)->properties & (zetscript::STACK_ELEMENT_PROPERTY_OBJECT|zetscript::STACK_ELEMENT_PROPERTY_OBJECT_IN_CONTAINER)) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_MEMBER_FUNCTION_SCRIPT_OBJECT))
+#define STACK_ELEMENT_IS_ZETOBJECT_SCRIPT_OBJECT(stk) 			(((stk)->properties & (zetscript::STACK_ELEMENT_PROPERTY_OBJECT|zetscript::STACK_ELEMENT_PROPERTY_OBJECT_IN_CONTAINER)) && (((zetscript::ScriptObject *)(stk)->value)->script_type_id==zetscript::SCRIPT_TYPE_ID_ZETSCRIPT_SCRIPT_OBJECT))
 
 #define ZS_GET_STK_VAR_REF(stk)  								(((VarRefObject *)((stk))->value)->getStackElementPtr())
 
@@ -76,10 +76,12 @@ namespace zetscript{
 		STACK_ELEMENT_PROPERTY_MEMBER_PROPERTY		=0x0100,
 		STACK_ELEMENT_PROPERTY_OBJECT				=0x0200,
 		STACK_ELEMENT_PROPERTY_CONTAINER_SLOT		=0x0400,
-		STACK_ELEMENT_PROPERTY_MAX					=0x0800,
+		STACK_ELEMENT_PROPERTY_OBJECT_IN_CONTAINER	=0x0800,
+		STACK_ELEMENT_PROPERTY_MAX					=0x1000,
 		//-- VM RUNTIME
 		STACK_ELEMENT_PROPERTY_PTR_STK				= STACK_ELEMENT_PROPERTY_MAX,	 				// 0x1000
-		STACK_ELEMENT_PROPERTY_READ_ONLY			=0x1000,
+		STACK_ELEMENT_PROPERTY_READ_ONLY			=0x2000,
+
 
 	}StackElementProperty;