+++ src/vm/vm_push_stk.h
@@ -37,7 +37,7 @@ data->vm_stk_current->value=ref; \
 data->vm_stk_current->properties=STACK_ELEMENT_PROPERTY_FUNCTION; \
 data->vm_stk_current++;
 
-#define ZS_VM_PUSH_STK_TYPE(ref) \
+#define ZS_VM_PUSH_STK_SCRIPT_TYPE(ref) \
 data->vm_stk_current->value=ref; \
 data->vm_stk_current->properties=STACK_ELEMENT_PROPERTY_SCRIPT_TYPE_ID; \
 data->vm_stk_current++;